allow the acceptance sub to return two values
We used to combine the boolean value of the result and the full content of the
errors all in one object, and use a boolean overload to access the former, but
it was determined that boolean overloads that can return false are problematic
in other areas, so instead we prefer just a boolean result from the validation
sub.. that value was still being jsonified, even though the error information
was no longer present, so we return it now as an (optional) second argument.

Callers can still provide their sub as:

  ...
  return $result->valid;

or switch to:

  ...
  return ($result->valid, $result);

...and this should work with either newer or older versions of TJSA.

And if you return a result that has a boolean overload, that will still work
too.